Return uintptr_t from vTableOrITableIndexFromMemberName()
The result is the value of J9JNIMethodID::vTableIndex, which is UDATA.
The high bit of this value is a flag that distinguishes VFT offsets from
iTable indices. The previous return type (int32_t) required this value
to be truncated on 64-bit platforms, losing the value of the flag,
leaving callers with no way to distinguish the cases.

+++ runtime/compiler/env/VMJ9.h
@@ -811,13 +811,13 @@ public:
     *    Return vtable or itable index of a method represented by MemberName
     *    Caller must acquire VM access
     */
-   virtual int32_t vTableOrITableIndexFromMemberName(uintptr_t memberName);
+   virtual uintptr_t vTableOrITableIndexFromMemberName(uintptr_t memberName);
    /*
     * \brief
     *    Return vtable or itable index of a method represented by MemberName
     *    VM access is not required
     */
-   virtual int32_t vTableOrITableIndexFromMemberName(TR::Compilation* comp, TR::KnownObjectTable::Index objIndex);
+   virtual uintptr_t vTableOrITableIndexFromMemberName(TR::Compilation* comp, TR::KnownObjectTable::Index objIndex);
    /*
     * \brief
     *    Create and return a resolved method from member name index of an invoke cache array.
